Adeed abstract class for BaseDeepNetworkPyTorch
