+++ sktime/networks/base.py
@@ -1,27 +1,6 @@
 """Abstract base class for deep learning networks."""
-
-__author__ = ["Withington", "TonyBagnall"]
-
-from abc import ABC, abstractmethod
-
-from sktime.base import BaseObject
+from abc import ABC
 
 
-class BaseDeepNetwork(BaseObject, ABC):
-    """Abstract base class for deep learning networks."""
-
-    @abstractmethod
-    def build_network(self, input_shape, **kwargs):
-        """Construct a network and return its input and output layers.
-
-        Parameters
-        ----------
-        input_shape : tuple
-            The shape of the data fed into the input layer
-
-        Returns
-        -------
-        input_layer : a keras layer
-        output_layer : a keras layer
-        """
-        ...
+class BaseDeepNetworkPyTorch(ABC):
+    """Abstract base class for deep learning networks using pytorch."""
